+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## v3.5.4
+### Changes
+* Add timing of shader parse to single shader or material instead of just for whole folder parses in the Shader Analyzer Debugger.
+
+### Bug Fixes
 * Fix first merged material blob on mesh generating all the animated property arrays instead of only the ones that are used by the meshes in the blob.
 
 ## v3.5.3